[Refactor] : User entity gender 타입 변경 (#3)
