--- src/main/java/com/potato/balbambalbam/entity/User.java
@@ -1,28 +1,0 @@
-package com.potato.balbambalbam.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@Table(name="user")
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private Integer age;
-
-    @Column(nullable = false)
-    private Byte gender; // 0 : 여성, 1 : 남성
-
-    @Column(nullable = false, unique = true)
-    private String email;
-}
