+++ src/main/java/com/potato/balbambalbam/user/controller/UserController.java
@@ -1,8 +1,8 @@
 package com.potato.balbambalbam.user.controller;
 
-import com.potato.balbambalbam.entity.User;
+import com.potato.balbambalbam.data.entity.User;
 import com.potato.balbambalbam.user.dto.UserDto;
-import com.potato.balbambalbam.user.repository.UserRepository;
+import com.potato.balbambalbam.data.repository.UserRepository;
 import com.potato.balbambalbam.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
